Return a non-zero exit code for all unhandled errors.
`main()` returns a result only for errors we can't easily intercept
and log. An example is config file parsing. The config file is the
source of truth for the kind of logger we need to build. Any `Err's
returned are displayed to STDERR and result in a non-zero exit
code. And due to the behavior of our logger, simply calling `exit()`
early prevents any error logging from reaching its destination.