do not attempt to reconnect on every packet received
If a gateway is connected, attempt to send all waiting packets.
Otherwise we store and move on, the reconnection logic will trigger
sending the packets when it is successful.

I changed the reconnect logging verbage to only be about the connection.
It's a bit confusing to read the logs and immediately see that you're
"reconnecting" when the gateway is connecting for the first time.