Don’t reconnect immediately on router failure
This change will recalculate the reconnect timer based on the fact that there was an error during the router.

If an error happens during a fresh connection the reconnect timer will be set to a short time using `next_connect` which will not try to connect immediately. The subseequent timer firing will end up reconnecting the router and sending the messages